refactor based on PR (#527)
