+++ apps/api/src/domain/domainModelValidators/errors/bibliographic-citation/bibliographic-citation-has-no-data.error.ts
