+++ apps/api/src/domain/models/bibliographic-citation/journal-article-bibliographic-citation/commands/create-journal-article-bibliographic-citation.command.integration.spec.ts
@@ -28,7 +28,7 @@ import AggregateIdAlreadyInUseError from '../../../shared/common-command-errors/
 import CommandExecutionError from '../../../shared/common-command-errors/CommandExecutionError';
 import InvalidExternalStateError from '../../../shared/common-command-errors/InvalidExternalStateError';
 import UuidNotGeneratedInternallyError from '../../../shared/common-command-errors/UuidNotGeneratedInternallyError';
-import { BibliographicCitationType } from '../../types/bibliogrpahic-citation-type';
+import { BibliographicCitationType } from '../../types/bibliographic-citation-type';
 import { CreateJournalArticleBibliographicCitation } from './create-journal-article-bibliographic-citation.command';
 
 const commandType = 'CREATE_JOURNAL_ARTICLE_BIBLIOGRAPHIC_CITATION';
