+++ apps/api/src/domain/models/bibliographic-citation/journal-article-bibliographic-citation/entities/journal-article-bibliographic-citation-data.entity.ts
@@ -5,7 +5,7 @@ import { isNullOrUndefined } from '../../../../utilities/validation/is-null-or-u
 import BaseDomainModel from '../../../BaseDomainModel';
 import BibliographicCitationCreator from '../../common/bibliographic-citation-creator.entity';
 import { BibliographicCitationDataUnionMember } from '../../shared/bibliographic-citation-union-data-member.decorator';
-import { BibliographicCitationType } from '../../types/bibliogrpahic-citation-type';
+import { BibliographicCitationType } from '../../types/bibliographic-citation-type';
 
 const isOptional = true;
