+++ apps/api/src/domain/models/bibliographic-citation/types/bibliographic-citation-type.ts
